backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "deepagents>=0.4.11",
    "fastapi>=0.135.1",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0",
    "langgraph-checkpoint-mongodb>=0.3.1",
    "langsmith>=0.7.22",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.15.5",
    "sqlalchemy[asyncio]>=2.0.48",
    "tavily-python>=0.7.23",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
